Rename dtypes.py to simulation_utils.py to reflect updates
