+++ utils/__init__.py
@@ -1,4 +1,12 @@
 # Import in __init__.py to make it available in the package
-from utils.dtypes import (
-    Strategy, Action, History, Player, battle, round_probabilistically, random_action, PAYOFF_MATRIX, STARTING_POPULATION
+from utils.simulation_utils import (
+    Strategy, 
+    Action, 
+    History, 
+    Player,
+    battle, 
+    round_probabilistically, 
+    random_action, 
+    PAYOFF_MATRIX, 
+    STARTING_POPULATION
 )
